Added Notifications.
Notifications will primarily be used to serve the user information message like "New Request Pending". The system can be expanded later on to a messaging system.

If you are a developer you can create message by going to notification/create . This has been done so the system can easily be tested. But will of course be visible in an easy-to-access location when the transition to a message system has been made.

Patch by: Lennard de Rijk