No override needed in notifications.py
The only downside is that the ?s=0 parameter will still be appended
but is not handled in list(). The right solution would be to update
list() so that it does say "Message Sent" or such, to give the user
an indication that their message was sent succesfully.

Patch by: Sverre Rabbelier