Edited the method used to clean link_id in params.py.
Removed the clean_new_link_id from the logic/cleaning.py module because this is not used anymore.

This method should be replaced by a method that is used after all the fields have been cleaned so that the actual key_name of an entity can be constructed.

Patch by: Lennard de Rijk