Added a new view to help with application -> group process.
The applicant view is available only when there is an accepted application present. This will be checked by the access checker in an upcoming patch. The view uses a special form so that certain fields cannot be edited by the user. These fields correspond to important fields filled in during the application process.

Create access for a club has now been set to Host level.

Patch by: Lennard de Rijk