Make the ToS text in-line and hook up thick-box in the help-text.
This way the user can see the ToS in scrollable form between the
link_id and the 'agree to tos' checkbox. Also, when clicking on the
link to the 'side-wide terms of service' a thickbox will pop up and
display the Terms of Service in an overlay.

Patch by: "Mario Ferraro" <fadinlight@gmail.com>