Stylefixes in access.py and removal of self.deny calls
The result of self.deny() is not very helpfull to the user, as such
we should not use it if we can provide a more sensible message.
Also made the document access checks use getForKeyFieldsOr404 to
prevent an error page when the specified document doesn't exist.

Patch by: Sverre Rabbelier