Redone the listing in group_app.py.
There is now a list_self that only shows the group application of the current user in the given a scope. The redirects point to edit or applicant when necessary.

Review_overview will show all the group applications that can be reviewed in the given scope, the redirect points to review for all statusses.

And list will just show all the group applications and redirect you to edit.

Patch by: Lennard de Rijk
Reviewed by: to-be-reviewed