Member template in organization is now called Contributor Template.
This matches the Organization Application, this way the information is automatically transferred to the new Organization.

Note that the template shows raw HTML, we might change the template tag to use |safe for it's input to show the true layout of the template. I've sent an email to SRabbelier about that.

Patch by: Lennard de Rijk
Reviewed by: to-be-reviewed