Major update to the views concerning StudentProjects.
Organization admins can now reassign a Mentor for each of their active (meaning student has not failed or finished) projects.

To facilitate this an overview has been added for the Organization Admins in which they can choose which Project to edit.

Students can now edit the public information that is used in their projects.

Several TODO's have been added to make it easier to navigate to these pages.

Patch by: Lennard de Rijk
Reviewed by: to-be-reviewed