Changed the Survey public view to give a preview of the Survey.
Also changed the access rights for deleting a Survey because of the possible invalid references problem.

Added two TODO's to the _public method. The first one involves enabling the SurveyForm to generate a form without passing along the Logic since it is only used to retrieve the Survey entity which we already have available. The second one involves a brief discussion about the getFields method's functionality and name.