Fixed showing the correct list of projects for taking a Grading Project Survey.
Note that when a user is a mentor for a lot of Organizations the amount of subqueries might exceed the allowed amount. A TODO has been added to fix this issue. However this will hopefully not be a problem that pops up soon.

Also some minor docstring addition to Project Survey view.