Added view for showing Survey Records.
The current access level is set to Developer only until this feature is complete
d.

A start has also been made on the viewResults method which will list the appropr
iate SurveyRecords for a given survey with a redirect to the viewRecord view.

Patch by: Lennard de Rijk and James Levy