Changed all status to start with Uppercase letters, made 3 properties required.
Changed all the status to start with Uppercase letters because they are
publicly visible. Also changed required=True to False for created_by,
modified_by property to allow site admin to create and modify tasks. Also
changed required=True to False for history to allow history not to be written
until the task is published.