Changed the working of the iterative_task decorator.
The decorator now needs the logic for the model which it fetches. This is much nicer then "fetching" it using import.

Also order has been removed since it has no use when iterating over all entities when key is involved. Fields to filter can now dynamically be set by the function that is wrapped.

Note that it will merge the filter with the task_default kwarg and it will only use the start_key if their is no key present in the context.

Also note that task_default gathers all kwargs. I tried to do it otherwise (with default kwargs) but I couldn't get it to work. Feel free to improve :).