getListContentForData function added to lists helper.
It allows to manually set a list of entities which is to be displayed on the view. Previously, the only function was getListContent, but it retrived data on its own by a single query. The new function can be useful whenever it is impossible or very awkward to obtain entities in such a way (for example more sophisticated SQL statements).

Additionally, the getListContent function is reconstructed so that it collects the data first and then calls getListContentForData.