Bug on GHOP task public view fixed.
When a student wanted to submit work, he had to select 'Submit work and Request for review' action. When the student submitted the comment with errors, the page was reloded, action option was still set to 'Submit work and Request for review', but the Submit Work input was not present. This commit fixes this issue.