Get current row data from row number instead from selected row.
Because we changed from selectRow to OnCellSelect to make the row
action not be triggered when clicking on the checkbox in multiselect
tables, the selected id coming from selrow gridparam could be
undefined (because no row could be selected if clicking on a cell in
a row which was already selected). So now we use the actual row number
that is passed to the onCellSelect jqgrid event.