Set correct row index to update cells.
setCell function can't take as input the same row_id that comes
from getRowData, because it starts from 1 instead that from 0.
In this way the first row was updated 2 times (one with 0 index
and one with 1 index), thus preventing the iteration continuing
till the last row (which remanined without spinner/checkbox).