?_field_hidden now hides the fields defined in ?_field_keys by default.
This ensures that the column models are only defined in 2 lists instead of 3, which makes it easier if for instance a list needs to be extended dynamically. Adding columns is then an append on the 2 lists on the SAME index rather then an append in one and figuring out where to position it in the ?_field_names param so it properly matches up.