Generalize enable/disable buttons logic to make it work onSelectAll.
Clicking on the header which selects all rows doesn't trigger an
onSelectRow action, but an onSelectAll action in jQGrid. This
action however, needs to be defined before jQGrid instantiation,
so it didn't work setting it afterwards using setGridParam.
So the logic to enable or disable buttons has been put into a
separate function and bound to onSelectRow and onSelectAll before
grid instantiation.

Fixes issue 728.