Middle click and CTRL+Left click now opens list items in a new tab.
Passing the event object along with the link in row actions now
we're able to detect the mouse button and key pressed while clicking
on rows. So now, by default, even if new window is not set in the
protocol configuration, middle clicks and CTRL+Left clicks open
row links in a new tab/window.

Fixes issue 740.